+++ packages/ripple-tide-api/types.d.ts
@@ -193,4 +193,12 @@ export interface IRplFeatureFlags {
    * @description Sets the theme variant of the site footer component
    */
   footerTheme?: 'neutral' | 'default'
+  /**
+   * @description Sets which search connector to use for content collection queries
+   */
+  contentCollectionSearchConnector?: 'appSearch' | 'elasticsearch'
+  /**
+   * @description Option to disable the display of topics and tags on all content types
+   */
+  disableTopicTags?: boolean
 }